reorganized structure to work with existing app setup
