+++ app.py
@@ -1,4 +1,4 @@
-from app.main import main
+from src.main import main
 
 if __name__ == "__main__":
     main()
